Moved test data into its own file.
